Refactor launcher controller
