+++ launcher/static/js/commons.js
@@ -3,11 +3,21 @@ function load_template(template_url){
     $.ajax({
         url: template_url,
         beforeSend: function() {
-            template_location.html("<i class='fa fa-circle-notch fa-spin'></i>");
+            template_location.html("<div class='my-2'><h5>Loading data <i class='fa fa-circle-notch fa-spin'></i></h5></div>");
+        },
+        error: function() {
+           template_location.html("<div class='my-2'>This service is temporary unavailable.</div>");
+           handle_route_button();
         },
         success: function(data) {
            template_location.html(data);
            handle_route_button();
+        },
+        complete: function() {
+           const spinner = $(".fa-sync", template_location.parent());
+            if (spinner.hasClass("fa-spin")){
+                spinner.removeClass("fa-spin");
+            }
         }
     });
 }
@@ -38,6 +48,15 @@ function handle_route_button(){
     });
 }
 
+function animate_sync(){
+    $(".fa-sync").click(function(){
+        if (!$(this).hasClass("fa-spin")){
+            $(this).addClass("fa-spin");
+        }
+    });
+}
+
 $(document).ready(function () {
     load_templates();
+    animate_sync();
 });
