+++ launcher/templates/bot_card.html
@@ -1,9 +1,3 @@
-<div class="card-header">
-    <h4>
-    OctoBot
-    </h4>
-</div>
-
 <!--Card content-->
 <div class="card-body grid">
     <div class="row col-12 mb-5">
