+++ launcher/templates/news_card.html
@@ -1,10 +1,1 @@
-<div class="card-header">
-    <h4>
-        OctoBot news feed
-    </h4>
-</div>
-
-<!--Card content-->
-<div class="card-body">
-    This service is temporary disabled.
-</div>
+This service is currently unavailable.
